requirements.txt
colorama==0.4.6
greenlet==3.2.1
numpy==2.2.5
pandas==2.2.3
pyodbc==5.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.40
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
